2.6. Changes [17.03.2018]

+++ sandbox/src/main/java/ru/stqa/pft/sandbox/Point.java
@@ -1,23 +1,9 @@
 package ru.stqa.pft.sandbox;
 
 public class Point {
-    public int x;
-    public int y;
+    public double x;
+    public double y;
 
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
